Added worker data, showed some visual indications when there are issues, added missing key props.
Logic for the issues -
The API returns a status field and a servers_issue field.
While I can assume that status would never be ok if the servers_issue is not null, I do not know for sure, so the status is shown regardless and the visual indication shows red when either status is not ok or servers_issue is not null.
This can result in "ok" being rendered as the status, but in red, because servers_issue is not null, but this is fine for now.