Added a socket manager.
It sends pings periodically and expects pongs soon after.
It makes sure the websocket is created and that it is open soon after.
It can also optionally make sure that data is received within a certain timeframe, given by the caller.
If any of those are not met, it triggers a failure notification.

Not used yet.