Added some tests for React.
Excluded the tests from tsconfig.vite.json to sidestep the unidentified Jest globals, but now it complains about the imports. Still, it runs, so it is better than nothing/before.
Added some paths to tsconfig.test.json to make the tests happier.
Added isolatedModules to make CSS imports work. It is deprecated, but when it is added to tsconfig itself, it breaks the tests completely, so the lesser of two evils. And added the identity-obj-proxy, which seals the deal for CSS imports.
Added the mocks as TypeScript files, otherwise TypeScript complains about tuple types in JSON as it infers them as union array types instead.
Added jest-dom for nicer assertions.
Made the backend types reflect, sometimes speculatively, the types of the status API.